+++ Changelog.md
@@ -6,6 +6,7 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 - Add support for variables attachment to query
 - Add support for variables passing query
+- Support for input objects variables
 
 ## 1.1: 2019-04-26